+++ include/utils/Folder.h
@@ -9,19 +9,30 @@
 #define _UTILS_FOLDER_H_
 
 #include "Exception.h"
-#include "utils/File.h"
 
 #include <functional>
 #include <memory>
 #include <string>
 #include <vector>
 
+class File {
+  public:
+    File(std::string file);
+
+    std::string& getName() { return name; }
+    std::string& getPath() { return path; }
+
+  private:
+    std::string name;
+    std::string path;
+};
+
 class Folder {
   public:
     Folder(std::string folder, bool listDotFiles = false);
 
-    std::string& getName();
-    std::string& getPath();
+    std::string& getName() { return name; }
+    std::string& getPath() { return path; }
 
     unsigned long fileCount();
     File& getFile(unsigned long i);
@@ -32,6 +43,7 @@ class Folder {
     Folder getParent();
 
     void executeRemoveFiles(std::function<bool (File& f)> func);
+    void findFilesEndingWith(std::vector<File>& found, std::string end, bool casesensitive = false);
 
     // Accessing a folder recursively
     // This treats all files in all subfolders as if they were in this folder
@@ -39,9 +51,9 @@ class Folder {
     void executeRemoveRecursiveFiles(std::function<bool (File& f)> func);
     std::string getRecursiveFileName(unsigned long i);
     File& getRecursiveFile(unsigned long i);
+    void findRecursiveFilesEndingWith(std::vector<File>& found, std::string end, bool casesensitive = false);
 
   private:
-
     void createFolderItems();
     int readFolderItems(std::vector<std::string>& foundFiles, std::vector<std::string>& foundFolders);