+++ include/utils/random.h
@@ -1,0 +1,14 @@
+/*!
+ * \file include/utils/random.h
+ * \brief Random number generation
+ *
+ * \author xythobuz
+ */
+
+#ifndef _UTILS_RANDOM_H_
+#define _UTILS_RANDOM_H_
+
+int randomInteger(int max, int min = 0);
+
+#endif
+